orchestrator/processing_coordinator/feedback_manager/gemini_generator/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytest>=7.0.0
